+++ src/index.ts
@@ -33,7 +33,13 @@ export {
     TokenRate
 } from "./interfaces/positions-contract"
 export { SRContractParams, SRContractSettings, SRPosition } from "./interfaces/savings-rate"
+export {
+    TokenswapContractParams,
+    TokenswapContractSettings,
+    TokenswapPositions
+} from "./interfaces/tokenswap"
 export { ITrxParams, ITrxParamsArgument } from "./interfaces/transaction"
 export { LiquidatorContract } from "./liquidator"
 export { PositionsContract } from "./main-positions"
 export { SavingsRateContract } from "./savings-rate"
+export { TokenSwapContract } from "./tokenswap"